The target is searched for pdb files and copied accordingly.

+++ AllureAttachmentWeaver/Program.cs
@@ -20,6 +20,13 @@ namespace AllureAttachmentWeaver
     {
         private static readonly ILog logger = LogManager.GetLogger("AllureAttachmentWeaverLogger");
         
+        // the PDB rewriting is based on this post:
+        // https://johnhmarks.wordpress.com/2011/01/19/getting-mono-cecil-to-rewrite-pdb-files-to-enable-debugging/
+        
+        /// <summary>
+        /// The entry point of the program, where the program control starts and ends.
+        /// </summary>
+        /// <param name="args">The command-line arguments.</param>
         public static void Main(string[] args)
         {
             Level loggingLevel = Level.Info;
@@ -48,7 +55,7 @@ namespace AllureAttachmentWeaver
             {
                 Weave(assembly, weaver);
 
-                Save(assembly);
+                Save(assembly, true);
             }
             catch(Exception e)
             {
@@ -77,33 +84,79 @@ namespace AllureAttachmentWeaver
             }
         }
 
-        private static void Save(AssemblyDefinition assembly)
+        private static void Save(AssemblyDefinition assembly, bool overWrite)
         {
             string[] args = Environment.GetCommandLineArgs();
 
             // already tested that atleast one argument is present.
             // user arguments array is 1 based.
-            string fileName = args[1];
+            string fullFileName = args[1];
 
-            string backup = GetFileNameWithKeyword(fileName, "original");
-            string generated = GetFileNameWithKeyword(fileName, "generated");
+            string target;
+            string pdb = GetMatchingPdb(fullFileName);
             
-            logger.Debug("Saving assembly to: " + fileName);
+            if (overWrite)
+            {
+                string backup = GetFileNameWithKeyword(fullFileName, "original");
+                
+                CleanMove(fullFileName, backup);
+                
+                if (pdb != null)
+                {
+                    CleanMove(pdb, Path.ChangeExtension(backup, "pdb"));
+                }
+                
+                target = fullFileName;
+            }
+            else
+            {
+                string generated = GetFileNameWithKeyword(fullFileName, "generated");
+                    
+                if (File.Exists(generated))
+                {
+                    File.Delete(generated);
+                }
+                    
+                if (pdb != null)
+                {
+                    string generatedPdb = Path.ChangeExtension(generated, "pdb");
+                 
+                    if (File.Exists(generatedPdb))
+                    {
+                        File.Delete(generatedPdb);
+                    }
+                }
+                    
+                target = generated;
+            }            
+            
+            logger.Debug("Saving assembly to: " + target);
 
-            if (File.Exists(generated))
-                File.Delete(generated);
-
-            if (File.Exists(backup))
-                File.Delete(backup);
-
-            assembly.Write(generated);
-
-            File.Move(fileName, backup);
-            File.Move(generated, fileName);
-
-            logger.Info("Assembly saved to: " + fileName);
+            assembly.Write(target);
+            
+            logger.Info("Assembly saved to: " + target);
         }
 
+        private static string GetMatchingPdb(string fullFileName)
+        {
+            string fullPdbFileName = Path.ChangeExtension(fullFileName, "pdb");
+            
+            if (File.Exists(fullPdbFileName))
+                return fullPdbFileName;
+            
+            return null;
+        }
+        
+        private static void CleanMove(string source, string destination)
+        {
+            if (File.Exists(destination))
+            {
+                File.Delete(destination);
+            }
+
+            File.Move(source, destination);
+        }
+        
         private static string GetFileNameWithKeyword(string filePath, string keyword)
         {
             string backup = Path.GetFileNameWithoutExtension(filePath) + "." + keyword + Path.GetExtension(filePath);
